python-service/requirements.txt
flask==3.0.0
discord.py==2.3.2
requests==2.31.0